check owner before update and delete jobs

+++ server/controllers/jobsController.js
@@ -5,6 +5,7 @@ const ErrorHandler = require('../utils/errorHandler')
 const catchAsyncErrors = require('../middlewares/catchAsyncErrors')
 const APIFilters = require('../utils/apiFilters')
 const path = require('path')
+const fs = require('fs')
 
 // get all jobs => /api/v1/jobs
 exports.getJobs = catchAsyncErrors(async (req, res, next) => {
@@ -65,6 +66,11 @@ exports.updateJob = catchAsyncErrors(async(req, res, next) => {
         return next(new ErrorHandler('Job not found', 404));
     }
 
+    // check if the user is owner
+    if(job.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorHandler(`User ${req.user.id} is not allowed to update this job`))
+    }
+
     job = await Job.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
         runValidators: true,
@@ -79,9 +85,24 @@ exports.updateJob = catchAsyncErrors(async(req, res, next) => {
 });
 
 exports.deleteJob = catchAsyncErrors(async (req, res, next) => {
-    let job = await Job.findById(req.params.id);
+    let job = await Job.findById(req.params.id).select('+applicantsApplied');
     if (!job) {
         return next(new ErrorHandler('Job not found', 404));
+    }
+
+    // check if the user is owner
+    if(job.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorHandler(`User ${req.user.id} is not allowed to delete this job`))
+    }
+
+    // deleting files associated with job
+    
+    for(let i = 0; i<delJob.applicantsApplied.length; i++) {
+        let filepath = `${__dirname}/public/uploads/${job.applicantsApplied[i].resume}`.replace('\\controllers', '');
+
+        fs.unlink(filepath, err => {
+            if (err) return console.log(err);
+        });
     }
 
     job = await Job.findByIdAndDelete(req.params.id);
